requirements.txt
dlib==19.24.1
numpy==1.24.2
opencv_contrib_python==4.7.0.72
pandas==1.5.3
Pillow==9.5.0
streamlit==1.22.0
